feat: projeto finalizado

+++ LiberadorSUAT/Entidades/Arquivo.cs
@@ -64,39 +64,42 @@ namespace LiberadorSUAT.Models
                 }
             }
         }
-        public void ExcluirArquivos(ListView listView, string tipo)
+        public void ExcluirArquivos(ListView listView, string tipo, CheckBox check)
         {
-            foreach (ListViewItem item in listView.Items)
+            if (listView.Items.Count > 0)
             {
-                string bover = item.SubItems.ToString();
-                if (item.Checked)
+                foreach (ListViewItem item in listView.Items)
                 {
-                    int index = item.Index;
-                    string nomeArquivo = listView.Items[index].SubItems[index+1].Text;
-
-                    switch (tipo)
+                    if (item.Checked)
                     {
-                        case "scripts":
-                            File.Delete(Directory.GetCurrentDirectory() + @"\scripts\" + Path.GetFileName(nomeArquivo));
-                            break;
+                        int index = item.Index;
+                        string nomeArquivo = listView.Items[index].SubItems[index + 1].Text;
 
-                        case "documentos":
-                            File.Delete(Directory.GetCurrentDirectory() + @"\documentacao\" + Path.GetFileName(nomeArquivo));
-                            break;
+                        switch (tipo)
+                        {
+                            case "scripts":
+                                File.Delete(Directory.GetCurrentDirectory() + @"\scripts\" + Path.GetFileName(nomeArquivo));
+                                break;
 
-                        case "sistema":
-                            File.Delete(Directory.GetCurrentDirectory() + @"\sistema\" + Path.GetFileName(nomeArquivo));
-                            break;
+                            case "documentos":
+                                File.Delete(Directory.GetCurrentDirectory() + @"\documentacao\" + Path.GetFileName(nomeArquivo));
+                                break;
+
+                            case "sistema":
+                                File.Delete(Directory.GetCurrentDirectory() + @"\sistema\" + Path.GetFileName(nomeArquivo));
+                                break;
+                        }
+
+                        listView.Items.RemoveAt(item.Index);
                     }
-                    
-                    listView.Items.RemoveAt(item.Index);
-                }
-                else
-                {
-                    MessageBox.Show("Nenhuma alteração foi selecionada.");
                 }
             }
+            else
+            {
+                MessageBox.Show("Nenhuma alteração foi selecionada.");
+            }
         }
+
         public void percorrerDiretorioArquivos(string caminho, string tipo)
         {
             string[] files = Directory.GetFiles(caminho, "*", SearchOption.AllDirectories);
